Clean up read/write ops
